Refactor trelloColors to add more colors later
